Integrated firebase for community group feature

+++ DalCommunity/app/src/main/java/com/csci5708/dalcommunity/model/ChatMessage.kt
@@ -1,3 +1,3 @@
 package com.csci5708.dalcommunity.model
 
-data class ChatMessage(val message:String,val sentByName:String, val sentById:String,val messageTime:Int)
+data class ChatMessage(val message:String,val sentByName:String, val sentById:String,val messageTime:Long)
